fix(sdk): address review findings
