+++ clients/typescript/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the TypeScript SDK are documented here.
 ## [0.3.0] - 2026-06-05
 
 ### Added
+
 - Initial public release of the TypeScript SDK
 - `Client` for authentication and sandbox CRUD
 - `Sandbox` helpers for sync, batch, and streaming exec
